Reestructuré archivos,se simula petición a /api/colors,  y cree Dockerfile
