+++ .dockerignore
@@ -1,0 +1,4 @@
+./.next
+./node_modules
+.gitignore
+README.md
