+++ .env
@@ -1,0 +1,1 @@
+NEXT_PUBLIC_IP=http://localhost:3000
