+++ docker-compose.yml
@@ -1,0 +1,11 @@
+version: "3"
+
+services:
+  front-colors:
+    build:
+      context: .
+      dockerfile: Dockerfile
+    container_name: front-colors
+    restart: always
+    ports:
+      - 4002:3000
