+++ src/constants/properties.js
