+++ src/pages/api/colors.js
@@ -1,3 +1,5 @@
+import bd_rios from "@/pages/api/BD-RIOS.json";
+
 export default function handler(req, res) {
-  res.status(200).json({});
+  res.status(200).json(bd_rios);
 }
